facerecognition-main/requirements.txt
boto3==1.34.80
botocore==1.34.80
click==8.1.7
dlib==19.24.4
face-recognition==1.3.0
face-recognition-models==0.3.0
jmespath==1.0.1
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
psycopg==3.1.18
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-decouple==3.8
RPi.GPIO==0.7.1
s3transfer==0.10.1
six==1.16.0
typing_extensions==4.11.0
urllib3==2.2.1
